+++ src/api/root.rs
@@ -13,7 +13,7 @@ pub async fn route() -> Json<Response> {
     let response = Response {
         url: "https://icebreakers.blockfrost.io".to_string(),
         version: env!("CARGO_PKG_VERSION").to_string(),
-        commit: env!("GIT_COMMIT_HASH"),
+        commit: env!("GIT_REVISION"),
         healthy: true,
     };