+++ src/main.rs
@@ -16,7 +16,6 @@ use clap::Parser;
 use colored::Colorize;
 use config::{Args, Config};
 use db::DB;
-use dotenvy;
 use std::net::SocketAddr;
 use tracing_subscriber::fmt::format::Format;